+++ crates/sncast/src/starknet_commands/multicall/invoke.rs
@@ -1,5 +1,6 @@
-use anyhow::{Context, Result};
+use anyhow::Result;
 use clap::Args;
+use sncast::helpers::configuration::CastConfig;
 use starknet_rust::core::{types::Call, utils::get_selector_from_name};
 
 use crate::{
@@ -22,8 +23,12 @@ pub struct MulticallInvoke {
 }
 
 impl MulticallInvoke {
-    pub fn new_from_item(item: &InvokeItem, contracts: &ContractRegistry) -> Result<Self> {
-        let calldata = parse_inputs(&item.inputs, contracts)?;
+    pub fn new_from_item(
+        item: &InvokeItem,
+        contracts: &ContractRegistry,
+        config: &CastConfig,
+    ) -> Result<Self> {
+        let calldata = parse_inputs(&item.inputs, contracts, config)?;
         let invoke = MulticallInvoke {
             common: InvokeCommonArgs {
                 contract_address: ContractAddress(item.contract_address.clone()),
@@ -37,17 +42,17 @@ impl MulticallInvoke {
         Ok(invoke)
     }
 
-    pub async fn build_call(&self, contract_registry: &mut ContractRegistry) -> Result<Call> {
+    pub async fn build_call(
+        &self,
+        contract_registry: &mut ContractRegistry,
+        config: &CastConfig,
+    ) -> Result<Call> {
         let selector = get_selector_from_name(&self.common.function)?;
-        let arguments = replaced_arguments(&self.common.arguments, contract_registry)?;
-        // TODO: add full support for multicall id <> alias id resolution (will be done in next PR)
-        let contract_address = if let Some(id) = self.common.contract_address.0.as_id() {
-            contract_registry
-                .get_address_by_id(id)
-                .with_context(|| format!("Failed to find contract address for id: {id}"))
-        } else {
-            self.common.contract_address.try_into_felt()
-        }?;
+        let arguments = replaced_arguments(&self.common.arguments, contract_registry, config)?;
+        let contract_address = self
+            .common
+            .contract_address
+            .resolve_in_multicall(contract_registry, config)?;
 
         let calldata = if let Some(raw_calldata) = &arguments.calldata {
             calldata_to_felts(raw_calldata)?